Divide `polling` into submodules
